refactor(blackjack): 인터페이스를 추가하여 골격 구현 방식으로 수정
